TP03
moved files
renames components
added styles
updated imports
