+++ src/App.js
@@ -1,10 +1,14 @@
+/**
+* Amin MOHAMED
+* Version 1.0
+*/
 import React from 'react';
-import './App.css';
 import { Footer, Header, Main } from './components';
+import './css/App.css';
 
 const App = () => {
   return (
-    <div className="app flex column justify-center fill-parent">
+    <div className="App flex column justify-center fill-parent">
       <Header />
       <Main />
       <Footer />
