+++ src/components/Footer.js
@@ -1,0 +1,17 @@
+/**
+* Amin MOHAMED
+* Version 1.0
+*/
+import React from 'react';
+import { Clock } from '.';
+import '../css/Footer.css';
+
+const Footer = () => {
+  return (
+    <footer className='Footer flex'>
+      <Clock /> 
+    </footer>
+  );
+};
+
+export default Footer;
