+++ src/components/Header.js
@@ -1,0 +1,17 @@
+/**
+* Amin MOHAMED
+* Version 1.0
+*/
+import React from 'react';
+import { Logo } from '.';
+import '../css/Header.css';
+
+const Header = () => {
+  return (
+    <header className='Header flex'>
+      <Logo /> 
+    </header>
+  );
+};
+
+export default Header;
