--- src/components/clock/Clock.js
@@ -1,13 +1,0 @@
-import React from "react";
-import './Clock.css';
-
-let TIME = new Date();
-
-const Clock = () => {
-  return (
-    <div className="clock flex center">{TIME.toUTCString()}</div>
-  );
-}
-
-
-export default Clock;
