--- src/components/footer/Footer.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Clock } from '..';
-import './Footer.css';
-
-const Footer = () => {
-  return (
-    <footer className='flex'>
-      <Clock /> 
-    </footer>
-  );
-};
-
-export default Footer;
