--- src/components/header/Header.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Logo } from '..';
-import './Header.css';
-
-const Header = () => {
-  return (
-    <header className='flex'>
-      <Logo /> 
-    </header>
-  );
-};
-
-export default Header;
