+++ src/components/index.js
@@ -1,6 +1,6 @@
-export { default as Clock } from './clock/Clock.js';
-export { default as Footer } from './footer/Footer.js';
-export { default as Header } from './header/Header.js';
-export { default as Logo } from './logo/Logo.js';
-export { default as Main } from './main/Main.js';
+export { default as Clock } from './Clock.js';
+export { default as Footer } from './Footer.js';
+export { default as Header } from './Header.js';
+export { default as Logo } from './Logo.js';
+export { default as Main } from './Main.js';
