+++ src/css/App.css
@@ -1,0 +1,33 @@
+.Digi-header {
+  background-color: #738fba;
+  font-family: Helvetica, Arial, sans-serif;
+  display: flex;
+  align-items: center;
+  padding: 5px;
+  border-radius: 5px;
+  color: #c03000;
+}
+
+.Digi-footer {
+  background-color: #bf2657;
+  font-family: Helvetica, Arial, sans-serif;
+  display: flex;
+  align-items: center;
+  padding: 7px;
+  border-radius: 3px;
+  color: #c03000;
+}
+
+.Digi-main {
+  background-color: #d7d7e0;
+  display: flex;
+  justify-content: center;
+  border-radius: 7px;
+  border-color: #9191a1;
+  color: #4e876e;
+  font-size: 12px;
+}
+
+h2, a {
+  color: #4BB5C1;
+}
