+++ src/css/Clock.css
@@ -1,3 +1,3 @@
-.logo {
+.Clock {
   padding: 0 25px;
 }
