+++ src/css/Footer.css
@@ -1,3 +1,3 @@
-header {
+Footer {
   height: 50px;
 }
