+++ src/css/Header.css
@@ -1,3 +1,3 @@
-footer {
+Header {
   height: 50px;
 }
