+++ src/css/Logo.css
@@ -1,3 +1,3 @@
-.clock {
+.Logo {
   padding: 0 25px;
 }
