+++ src/index.js
@@ -1,3 +1,7 @@
+/**
+* Amin MOHAMED
+* Version 1.0
+*/
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
